requirements.txt
anyio==3.6.2
arrow==1.2.3
attrs==22.2.0
binaryornot==0.4.4
briefcase==0.3.13
build==0.10.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
gbulb==0.6.4
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
psutil==5.9.4
pycairo==1.23.0
Pygments==2.14.0
PyGObject==3.42.2
pyproject_hooks==1.0.0
pytest==7.2.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.6
text-unidecode==1.3
toga-core==0.3.0
toga-gtk==0.3.0
tomli_w==1.0.0
travertino==0.1.3
typing_extensions==4.5.0
urllib3==1.26.15
